requirements.prod.txt
fastapi==0.115.12
uvicorn==0.34.2
numpy==2.2.5
pillow==11.2.1
pillow-heif==0.16.0
python-multipart==0.0.20
python-magic==0.4.27
torch==2.6.0
torchvision==0.21.0
streamlit==1.44.1 